+++ packages/design-system/blocks/navigation/data.js
@@ -49,8 +49,17 @@ const data = [
     href: "/kontakt",
   },
   {
-    title: "Pliki do pobrania",
-    href: "/pliki-do-pobrania",
+    title: "Dokumenty",
+    links: [
+      {
+        title: "Umowy",
+        href: "/umowy",
+      },
+      {
+        title: "Ankiety",
+        href: "/ankiety",
+      },
+    ],
   },
 ]
