--- packages/frontend/app/pages/pliki-do-pobrania.jsx
@@ -1,5 +1,0 @@
-import FilesToDownloadPage from "../templates/FilesToDownload"
-
-const FilesToDownload = () => <FilesToDownloadPage />
-
-export default FilesToDownload
