+++ packages/frontend/app/pages/umowy.jsx
@@ -1,0 +1,5 @@
+import AgreementsPage from "../templates/Agreements"
+
+const Agreements = () => <AgreementsPage />
+
+export default Agreements
