+++ packages/frontend/app/templates/Agreements/components/index.js
