+++ packages/frontend/app/templates/Agreements/index.js
@@ -1,0 +1,3 @@
+import Agreements from "./Agreements"
+
+export default Agreements
