+++ packages/frontend/app/templates/Documents/components/DocumentsTemplate.jsx
@@ -5,34 +5,18 @@ import Typography from "design-system/components/typography"
 import List from "design-system/components/list"
 import ListItem from "design-system/components/listItem"
 
-import AdoptionSurvey from "frontend/public/documents/AdoptionSurvey.svg"
-import TempHomeSurvey from "frontend/public/documents/TempHomeSurvey.svg"
-import globalData from "shared/data"
-
 import * as Styled from "./DocumentsTemplate.styled"
 
-const list = [
-  "pobierz plik",
-  "wypełnij na komputerze lub wydrukuj",
-  `wypełnione dokumenty prześlij na adres: 
-    <a href="mailto:${globalData.contact.mail.text}">
-      ${globalData.contact.mail.text}
-    </a>
-  `,
-  "poczekaj na odpowiedź od nas",
-]
+const DocumentsTemplate = ({ data }) =>
+  data.map((document) => (
+    <Styled.SurveyContainer key={document.heading}>
+      <Styled.Image>{document.asset}</Styled.Image>
 
-const DocumentsTemplate = () => (
-  <>
-    <Styled.SurveyContainer>
-      <Styled.Image>
-        <AdoptionSurvey />
-      </Styled.Image>
       <TextBanner
-        heading="Ankieta przedadopcyjna"
-        size="small"
+        heading={document.heading}
         tabletLayout="left"
         mobileLayout="left"
+        size="small"
         button={
           <Button
             text="Pobierz"
@@ -40,15 +24,15 @@ const DocumentsTemplate = () => (
             color="primary"
             iconEnd="download"
             size="medium"
-            href={globalData.surveys.preAdoption}
+            href={document.link}
             target="_blank"
           />
         }
       >
-        <Typography variant="bodySmall">Co musisz zrobić?</Typography>
+        <Typography variant="bodyTitle">Co musisz zrobić?</Typography>
 
         <List gap={0}>
-          {list.map((item) => (
+          {document.steps.map((item) => (
             <ListItem key={item} iconName="dot" variant="gray">
               <div dangerouslySetInnerHTML={{ __html: item }} />
             </ListItem>
@@ -56,39 +40,6 @@ const DocumentsTemplate = () => (
         </List>
       </TextBanner>
     </Styled.SurveyContainer>
-    <Styled.SurveyContainer>
-      <Styled.Image>
-        <TempHomeSurvey />
-      </Styled.Image>
-      <TextBanner
-        heading="Ankieta dla domu tymczasowego"
-        size="small"
-        tabletLayout="left"
-        mobileLayout="left"
-        button={
-          <Button
-            text="Pobierz"
-            variant="fill"
-            color="primary"
-            iconEnd="download"
-            size="medium"
-            href={globalData.surveys.temporaryHome}
-            target="_blank"
-          />
-        }
-      >
-        <Typography variant="bodySmall">Co musisz zrobić?</Typography>
-
-        <List gap={0}>
-          {list.map((item) => (
-            <ListItem key={item} iconName="dot" variant="gray">
-              <div dangerouslySetInnerHTML={{ __html: item }} />
-            </ListItem>
-          ))}
-        </List>
-      </TextBanner>
-    </Styled.SurveyContainer>
-  </>
-)
+  ))
 
 export default DocumentsTemplate
