--- packages/frontend/app/templates/FilesToDownload/index.js
@@ -1,3 +1,0 @@
-import FilesToDownload from "./FilesToDownload"
-
-export default FilesToDownload
